Implementado nuevo endpoint para tareas pendientes por estudiante y optimizar consulta de la base de datos

+++ backend-api/src/main/java/org/example/backendapi/dto/AulaRequestDTO.java
@@ -9,5 +9,5 @@ public class AulaRequestDTO {
     private String nombre;
 
     @NotBlank(message = "El ID del profesor es obligatorio")
-    private Integer profesorId;
+    private Long profesorId;
 }
