+++ backend-api/src/main/java/org/example/backendapi/dto/PruebaResponseDTO.java
@@ -5,12 +5,12 @@ import java.time.Instant;
 
 @Data
 public class PruebaResponseDTO {
-    private Integer id;
+    private Long id;
     private String titulo;
     private String tipo;
     private String contenido;
     private Integer puntuacionMaxima;
     private Instant fechaCreacion;
     private Instant fechaLimite;
-    private Integer aulaId;
+    private Long aulaId;
 }
