+++ backend-api/src/main/java/org/example/backendapi/dto/UsuarioResponseDTO.java
@@ -5,11 +5,11 @@ import java.time.Instant;
 
 @Data
 public class UsuarioResponseDTO {
-    private Integer id;
+    private Long id;
     private String nombreUsuario;
     private String rol;
     private Instant fechaCreacion;
-    private Integer aulaId;
+    private Long aulaId;
     private Integer nivelActual;
     private Integer experienciaActual;
 }
